refactor(lease): extract shared cache_credentials and TempEnvGuard
- Move TempEnvGuard from exec.rs to lease.rs as a public type
- Extract duplicated credential caching logic into lease::cache_credentials()
- Add TempEnvGuard to interactive lease create for env var cleanup
- Both exec.rs and lease.rs now use the shared implementations
